+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 ### Deprecated
 - Changed `Enums` names to follow CapWords convention. Old names are still supported, but will soon be deprecated.
 
+### Changed
+- Change dictionary registers groups to objects
+- Parse CanOpen object type from xdf v3 dictionaries
+
 ### Added
 - Created ConfigurationFile class, used in load_configuration, save_configuration and check_configuration